backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pytest==7.4.3
httpx==0.25.2
psycopg2-binary==2.9.10
python-dotenv==1.0.0
